Move helper method from user views to profile views integration test
